Add desktop navigation
