+++ src/js/index.js
@@ -1,7 +1,6 @@
 // Component imports
 import { pageTitle } from './components/pageTitle';
 import defaultComp from './components/default';
-import mainNavigation from './components/navigation';
 import { headerSection } from './components/header';
 import { initialMessageComp } from './components/initialMessage';
 import { pageData } from '../data/data';
@@ -13,7 +12,6 @@ import { pageData } from '../data/data';
 const components = {
   default: defaultComp,
   pageTitle: pageTitle,
-  navigation: mainNavigation,
   header: headerSection,
   textBlockComp: initialMessageComp,
 };
